+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/utils/MessagesResending.kt
@@ -15,6 +15,7 @@ import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.executeUnsafe
 
 private typealias ChatIdMessageIdPair = Pair<ChatId, MessageIdentifier>
 
+@Deprecated("Deprecated for the reason of less obviously of which forwarded message where from forwarded")
 suspend fun cacheMessages(
     executor: RequestsExecutor,
     sourceChatId: ChatId,
@@ -37,6 +38,42 @@ suspend fun cacheMessages(
             messagesToDelete.add(it.chat.id to it.messageId)
         } as? AbleToBeForwardedMessage
     }.also {
+        if (clear) {
+            messagesToDelete.forEach {
+                executor.executeAsync(
+                    DeleteMessage(
+                        it.first,
+                        it.second
+                    )
+                )
+            }
+        }
+    }
+}
+
+suspend fun cacheMessagesToMap(
+    executor: RequestsExecutor,
+    sourceChatId: ChatId,
+    cacheChatId: ChatId,
+    messagesIds: Iterable<MessageIdentifier>,
+    clear: Boolean = true
+): Map<MessageIdentifier, AbleToBeForwardedMessage> {
+    val messagesToDelete = mutableListOf<ChatIdMessageIdPair>()
+
+    return messagesIds.mapNotNull { id ->
+        executor.executeUnsafe(
+            ForwardMessage(
+                sourceChatId,
+                cacheChatId,
+                id,
+                disableNotification = true
+            ),
+            retries = 3
+        ) ?.asMessage  ?.let {
+            messagesToDelete.add(it.chat.id to it.messageId)
+            id to it as AbleToBeForwardedMessage
+        }
+    }.toMap().also {
         if (clear) {
             messagesToDelete.forEach {
                 executor.executeAsync(